requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1
sqlparse==0.4.2
